+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # main [(unreleased)](https://github.com/whitesmith/rubycritic/compare/v4.8.0...main)
 
 * [BUGFIX] Fix sort and filters on the coverage page (by [@kcamcam][])
+* [CHANGE] Scope churn calculation to provided paths (by [@Fito][])
 
 # v4.8.0 / 2023-05-12 [(commits)](https://github.com/whitesmith/rubycritic/compare/v4.7.0...v4.8.0)